[refactor/#32]: LoanGoodsService 내부 함수추출로 코드 리팩토링
